+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/components/MappingWizardDatastoresStep/components/DatastoresStepForm/__tests__/DatastoresStepForm.test.js
@@ -9,7 +9,7 @@ import {
   targetClusters,
   sourceDatastores,
   targetDatastores
-} from './fixtures';
+} from '../fixtures';
 
 let selectedCluster;
 let targetCluster;
